Mark libpng as double_loadable
libpng is explicitly marked as double_loadable since it is one of the
(indirect) dependencies of the LLNDK library libnativewindow
(libft2 -> libpng)
and at the same time the lib itself is marked as VNDK. Such lib can be
double loaded inside a vendor process.

Note: even without this change, the library is already capable of being
double loaded due to the dependency graph around it. This change is to
make it explicit so that double loading of a library is carefully
tracked and signed-off by the owner of the lib.

Bug: 77155589
Test: m -j

Change-Id: I9668ec31a1aca0305bd7bd7c0687d2ad331e221d