Change required for adding libpdfium in pdfViewer app
We are creating a new component "pdf" inside MediaProvider and adding "libpdfium" as a dependency by static linking. As "libpdfium" is depending on "libft2" which is depending on "libpng",we need this change to build the module.
See the doc below for more details about the build issues without this change - https://docs.google.com/document/d/1-pRvjnvdRbq4GaF40w1862DcB-7RwHpUknvO5LVvjv4/edit?usp=sharing

Bug: 307868581
Test: Build the code
Change-Id: I11b907ab99ab722aa8436d1d23188e07851654ed